+++ demo/NestedGlobalScope/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { createMultiGlobalScopes } from '../../src/GlobalState/GlobalState';
 import { GlobalScope } from '../../src/GlobalState/GlobalScope';
-import { GlobalReducer } from '../../src/GlobalState/Reducer';
+import { GlobalReducer } from '../../src/GlobalState/GlobalReducer';
 import App from './App';
 
 const nestedScope = {