+++ src/index.ts
@@ -11,4 +11,4 @@ export {
 
 export { Scope } from './GlobalState/Scope';
 export { GlobalScope } from './GlobalState/GlobalScope';
-export { GlobalReducer } from './GlobalState/Reducer';
+export { GlobalReducer } from './GlobalState/GlobalReducer';